Strip out all recording logic
