+++ src-tauri/src/lib.rs
@@ -90,14 +90,7 @@ struct AppState {
   input_event_tx: broadcast::Sender<rdev::Event>,
   // Recording related
   is_recording: bool,
-  pause_recording_tx: Option<broadcast::Sender<()>>,
-  stop_recording_tx: Option<broadcast::Sender<()>>,
-  stop_barrier: Option<Arc<std::sync::Barrier>>,
   recording_manifest: Option<RecordingManifest>,
-  start_writing: Arc<AtomicBool>,
-  screen_recording_metadata: Option<ScreenRecordingMetadata>,
-  screen_frame_tx: Option<broadcast::Sender<Vec<u8>>>,
-  screen_segments: Vec<PathBuf>,
   // Editing related
   is_editing: bool,
   export_process: Option<Arc<Mutex<FfmpegChild>>>,
@@ -193,14 +186,7 @@ pub fn run() {
       magnifier_running: Arc::new(AtomicBool::new(false)),
       input_event_tx: input_event_tx.clone(),
       is_recording: false,
-      pause_recording_tx: None,
-      stop_recording_tx: None,
-      stop_barrier: None,
       recording_manifest: None,
-      start_writing: Arc::new(AtomicBool::new(false)),
-      screen_recording_metadata: None,
-      screen_frame_tx: None,
-      screen_segments: Vec::new(),
       is_editing: false,
       export_process: None,
     }))
