+++ src-tauri/src/recording/mod.rs
@@ -1,3 +1,8 @@
+mod audio;
+mod camera;
 pub mod commands;
+mod ffmpeg;
+mod file;
+mod input_events;
 pub mod models;
-mod service;
+mod screen;
