+++ src-tauri/src/recording/models.rs
@@ -1,25 +1,14 @@
-use std::{
-  path::{Path, PathBuf},
-  sync::{atomic::AtomicBool, Arc},
-};
+use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};
 use strum_macros::{AsRefStr, Display, EnumString};
 use tauri::{LogicalPosition, LogicalSize};
-use uuid::Uuid;
 
 #[derive(Debug, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Region {
   pub position: LogicalPosition<f64>,
   pub size: LogicalSize<f64>,
-}
-
-#[derive(Debug, Clone)]
-pub struct StreamSynchronization {
-  pub start_writing: Arc<AtomicBool>,
-  pub stop_tx: tokio::sync::broadcast::Sender<()>,
-  pub stop_barrier: Arc<std::sync::Barrier>,
 }
 
 #[derive(EnumString, AsRefStr, Display, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -61,33 +50,6 @@ pub enum RecordingFile {
   #[strum(serialize = "metadata.json")]
   #[serde(rename = "metadata.json")]
   Metadata,
-}
-
-impl RecordingFile {
-  fn base_name(&self) -> &'static str {
-    match self {
-      RecordingFile::Screen => "screen",
-      _ => todo!(),
-    }
-  }
-
-  fn extension(&self) -> &'static str {
-    match self {
-      RecordingFile::Screen => "mp4",
-      _ => todo!(),
-    }
-  }
-
-  /// Append uuid to name and create full path
-  pub fn segment_path(&self, dir: &Path) -> PathBuf {
-    let uuid = Uuid::new_v4();
-    dir.join(format!(
-      "{}-{}.{}",
-      self.base_name(),
-      uuid,
-      self.extension()
-    ))
-  }
 }
 
 #[derive(Debug, Clone, Serialize)]
